requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
groq
python-dotenv
